drill: 4.4

+++ matrix/books/Programming_Principles_CPP/ch4-computation/drills/main.cpp
@@ -2,8 +2,8 @@
 
 int main()
 {
-    int num1{0};
-    int num2{0};
+    double num1{0};
+    double num2{0};
 
     while (cin >> num1 >> num2)
     {
